requirements/base.txt
requests==2.8.1
futures==3.0.3
requests-futures==0.9.5
lxml==3.4.4
cssselect==0.9.1
gunicorn==19.4.1
bottle==0.12.9
PyMySQL==0.6.7
python-editor==0.4
Mako==1.0.2
MarkupSafe==0.23
SQLAlchemy==1.0.9
alembic==0.8.3
smtpapi==0.2.0
sendgrid==1.5.17
PyGithub==1.26.0
